02. Status Management
In this class, you learned about:
Know what state management is and what its importance in the video is the concept of state management;
Understanding how the only source of truth can assist in the organization and consistency of data when we manage states in the video the concept of state management;
Know how Flutter manages states natively through inherited widget and, in turn, how data is passed inside the inherited widget to control the status in the Inherited Widget video;
Understand the importance of inherited widget and also that its complexity in practice is not recommended for everyday use, thus introducing the provider - a state management library - in the video installing the provider;
Implement the provider in our application through the ChangeNotifierProvider and the models of the project in the video using the provider.
We completed class 2!