03. Consuming Data
In this class, you learned to:
Understand what the widget is consumer and its relevance when consulting the state directly from inside the widget tree with the video consumer;
Add customers to the customer list being managed by the provider through the consumer in the video add customers;
Read the customer type data contained in the model in the video by reading customer types;
Apply and understand how the Provider.of to access the state of a provider without this being directly inside the widget tree in the video types of customers.
We completed class 3!