04. Improving Knowledge
In this class, you learned to:
How to improve the implementation of our code, assigning the model the responsibility to edit data in the video "Improving the code";
Understand how state management works through the BloC standard in the "Knowing BloC" video";
Understand how state management works through Redux in the video "Discovering Redux";
To know better that there are other state managers besides those seen in class and that there is no better or worse; there is always what best suits what you need on a specific occasion, in the activity to learn more.
We completed class 4! Congratulations on getting here.